refactor: 결과 검색 키워드 클래스명 변경, 패키지 이동
- 다른 변수명, 메서드명과 통일하여 FindCommand 대신 SearchKeyword로 변경
- 컨트롤러에서만 사용되므로 패키지 이도
