+++ README.md
@@ -169,12 +169,15 @@
 #### 2단계 리뷰 반영 리팩터링 목록
 
 - [x] 모호한 네이밍 수정 (Result -> Item)
-- [ ] int값으로 주고받는 index를 의미있는 클래스로 원시값 포장하기
-    - [ ] index 검증 로직 유틸리티 클래스 대신 해당 클래스에서 직접 하기
+- [ ] index 원시값 포장하기
+    - [ ] index 검증 로직 유틸리티 클래스 대신 해당 클래스에서 직접 하기 => 범위를 모르는데 할 수가 있나?
     - [ ] index 검증 로직 부정확한 예외 메시지 수정
     - [ ] index 검증에 대한 커스텀 예외 검토
-- [ ] keyword 대소문자 구분 없이 모두 처리 가능하게 만들기
-- [ ] FindCommand enum에서 해당하는 상수 확인하게 만들기
+- [ ] 결과 검색 키워드 관련
+    - [x] 다른 변수명, 메서드명과 통일하여 클래스명 SearchKeyword로 변경
+    - [x] 컨트롤러에서만 사용되므로 패키지 이동
+    - [ ] keyword 대소문자 구분 없이 모두 처리 가능하게 만들기
+    - [ ] enum 클래스에서 해당하는 상수 확인하게 만들기
 - [ ] GameResult 클래스의 key, value도 원시값 포장하기
 - [ ] 부적절한 스트림 활용 수정하기
     - [ ] map 생성 시 스트림만으로 처리하거나 일반 foreach 사용하기 (학습 필요)
