+++ src/main/java/laddergame/controller/SearchKeyword.java
@@ -1,6 +1,6 @@
-package laddergame.view.constant;
+package laddergame.controller;
 
-public class FindCommand {
+public class SearchKeyword {
 
     public static final String FIND_ALL = "all";
     public static final String QUIT = "Q";
